break out UriViewer, deleting
